Add .gitignore, requirements.txt, and main.py
A new .gitignore file has been created to ignore unnecessary files and directories from the JetBrains IDEs, Python, and the venv. A requirements.txt file is added with a list of necessary python packages for the project. The main.py file has been added to the app directory, currently containing two basic routes for example purposes.