Add exception handler to app instantiation
The exception handler has been moved from the bottom of the main.py file to immediately after the FastAPI app is created. This is to ensure that the exception handler is added to the app as soon as it's created, improving error handling throughout the application. As a result, the startup and shutdown event handlers have been removed.